Feat. add useGetForeignDocuments
