+++ src/atoms/atoms.js
@@ -4,6 +4,7 @@ export const userAtom = atom("");
 
 export const databasesAtom = atom([]);
 export const documentsAtom = atom([]);
+export const foreignDocumentsAtom = atom([]);
 
 export const currentDBIdAtom = atom("");
 export const currentDBNameAtom = atom("");
